pyproject.toml
[tool.poetry]
name = "fastapi_zero"
version = "0.1.0"
description = ""
authors = ["Douglasgls <douglaspaz955@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
ice = "^0.0.2"
fastapi = {extras = ["standard"], version = "^0.114.2"}
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
alembic = "^1.13.2"
faker = "^29.0.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.10"
pyjwt = "^2.9.0"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix'
run = 'fastapi dev fast_zero/app.py --host 0.0.0.0 --port 7070'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

